Comply with XNB format (uint32 vs int32)
although Monogame use ints internally at some places, it's best to read
data as uint as defined in the XNB format, any overflow will be caught
when building with `checked` enabled.

+++ MonoGame.Framework/Content/ContentReaders/EffectReader.cs
@@ -16,7 +16,7 @@ namespace Microsoft.Xna.Framework.Content
         {
             var effect = existingInstance;
             
-            int dataSize = input.ReadInt32();
+            int dataSize = (int)input.ReadUInt32();
             byte[] data = input.ContentManager.GetScratchBuffer(dataSize);
             input.Read(data, 0, dataSize);
